+++ com.ibm.streamsx.inet/impl/java/src/com/ibm/streamsx/inet/http/HTTPRequestOperAPI.java
@@ -119,6 +119,7 @@ public class HTTPRequestOperAPI extends AbstractOperator {
     private ContentType                   contentTypeToUse = null;
     private ArrayList<String>             extraHeaders = null;
     private TupleAttribute<Tuple, String> extraHeaderAttribute = null;
+    private TupleAttribute<Tuple, String> accessTokenAttribute = null;
     
     //request configuration
     private TupleAttribute<Tuple, String> requestBodyAttribute = null;  //request body
@@ -210,6 +211,12 @@ public class HTTPRequestOperAPI extends AbstractOperator {
     public void setExtraHeaderAttribute(TupleAttribute<Tuple, String> extraHeaderAttribute) {
         this.extraHeaderAttribute = extraHeaderAttribute;
     }
+    @Parameter(optional = true, description = "Attribute that specifies the access token if `authenticationType` is `OAUTH2`. Any "
+            + "value of this attribute overwrites the value of the `accessToken` property in `authenticationFile` or in "
+            + "`authenticationProperties`. This parameter is meaningless if the `authenticationType` is different from `OAUTH2`")
+    public void setAccessTokenAttribute(TupleAttribute<Tuple, String> accessTokenAttribute) {
+        this.accessTokenAttribute = accessTokenAttribute;
+    }
     
     /***************************************
      * request configuration
@@ -218,8 +225,8 @@ public class HTTPRequestOperAPI extends AbstractOperator {
             + "these attributes is sent as request body in method POST. If parameter `requestAttributesAsUrlArguments` is true, "
             + "the request attributes are additionally appended as arguments to the url in method GET. "
             + "If this parameter is missing, "
-            + "all attributes, excluding those that are used to specify the URL, method, content type, Request url arguments "
-            + "or request attributes, are used in the request body. "
+            + "all attributes, excluding those that are used to specify the URL, method, content type, Request url arguments, "
+            + "extra header, access token or request body, are used in the request body. "
             + "One empty element defines an empty list which means no attributes are considered request attributes. ")
     public void setRequestAttributes(String[] requestAttributes) {
         // Is set in initialize with special treatment
@@ -502,6 +509,10 @@ public class HTTPRequestOperAPI extends AbstractOperator {
                 requestAttributes.remove(requestUrlArgumentsAttribute.getAttribute().getName());
             if (requestBodyAttribute != null)
                 requestAttributes.remove(requestBodyAttribute.getAttribute().getName());
+            if (extraHeaderAttribute != null)
+                requestAttributes.remove(extraHeaderAttribute.getAttribute().getName());
+            if (accessTokenAttribute != null)
+                requestAttributes.remove(accessTokenAttribute.getAttribute().getName());
         }
         //Check whether all attributes are string type if fixed content type is ne json
         /*if (((fixedContentType != null) && (contentTypeToUse != ContentType.APPLICATION_JSON)) || requestAttributesAsUrlArguments) {
@@ -667,6 +678,7 @@ public class HTTPRequestOperAPI extends AbstractOperator {
     //getter
     protected ArrayList<String>             getExtraHeaders() { return extraHeaders; }
     protected TupleAttribute<Tuple, String> getExtraHeaderAttribute() { return extraHeaderAttribute; }
+    protected TupleAttribute<Tuple, String> getAccessTokenAttribute() { return accessTokenAttribute; }
     
     //request config
     protected TupleAttribute<Tuple, String> getRequestBodyAttribute() { return requestBodyAttribute; }